Methods list

ImageGeoms.ImageGeom

ImageGeoms.ImageGeom

ImageGeoms.ImageGeom

ImageGeoms.ImageGeom

Base.show

ImageGeoms.downsample

ImageGeoms.downsample

ImageGeoms.downsample!

ImageGeoms.ellipse

ImageGeoms.embed

ImageGeoms.embed

ImageGeoms.embed!

ImageGeoms.expand_nz

ImageGeoms.getindex!

ImageGeoms.mask_or

ImageGeoms.maskit

ImageGeoms.oversample

ImageGeoms.plot

ImageGeoms.upsample

ImageGeoms.upsample!

Methods usage

ImageGeoms.ImageGeom — Type

ImageGeom{D,S}

Image geometry struct with essential grid parameters.

dims::Dims{D} image dimensions

deltas::S where S <: NTuple{D} pixel sizes, where each Δ is usually Real or Unitful.Length

offsets::NTuple{D,Float32} unitless

mask::AbstractArray{Bool,D} logical mask, often FillArrays.Trues(dims).

ImageGeoms.ImageGeom — Method

ig = ImageGeom(masktype::Mask ; kwargs...)

Constructor with specified mask, e.g., ImageGeom(MaskCircle() ; dims=(6,8)).

ImageGeoms.ImageGeom — Method

ig = ImageGeom(dims::Dims, deltas, offsets, [, mask])

Constructor for ImageGeom. The deltas elements should each be Real or a Unitful.Length. The dims, deltas and offsets tuples must be same length. Default mask is FillArrays.Trues(dims) which is akin to trues(dims).

ImageGeoms.ImageGeom — Method

ig = ImageGeom( ; dims=(nx,ny), deltas=(1,1), offsets=(0,0), mask=Trues )

Constructor using named keywords.

Base.show — Method

show(io::IO, ::MIME"text/plain", ig::ImageGeom)

ImageGeoms.downsample! — Method

downsample!(y, x, down::NTuple{Int})

Non-allocating version of y = downsample(x, down). Requires size(y) = ceil(size(x) / down).

ImageGeoms.downsample — Method

y = downsample(x, down::NTuple{Int})

Down-sample a Bool array x by "max pooling" in all dimensions. Output size will be ceil(size(x) / down). Thus, in a typical case where size(x) is integer multiple of down, the output size will be size(x) ÷ down.

ImageGeoms.downsample — Method

ig_down = downsample(ig, down::Tuple{Int})

Down sample an image geometry by the factor down. cf image_geom_downsample

ImageGeoms.ellipse — Method

ellipse(ig::ImageGeom{2} ; ...)

Ellipse that just inscribes the rectangle, but keeping a 1 pixel border due to some regularizer limitations.

ImageGeoms.embed! — Method

embed!(array, v, mask ; filler=0)

embed vector v of length sum(mask) into elements of array where mask is true, setting the remaining elements to filler (default 0).

ImageGeoms.embed — Method

array = embed(matrix::AbstractMatrix{<:Number}, mask::AbstractArray{Bool})

Embed each column of matrix into mask then cat along next dimension In:

matrix [sum(mask) L]

mask [(N)]

Out:

array [(N) L]

ImageGeoms.embed — Method

array = embed(v, mask ; filler=0)

embed vector v of length sum(mask) into elements of an array where mask is true; see embed!.

ImageGeoms.expand_nz — Method

ig_new = expand_nz(ig::ImageGeom{3}, nz_pad::Int)

Pad both ends along z.

ImageGeoms.getindex! — Method

getindex!(y::AbstractVector, x::AbstractArray{T,D}, mask::AbstractArray{Bool,D})

Equivalent to the in-place y .= x[mask] but is non-allocating.

For non-Boolean indexing, just use @views y .= A[index], per https://discourse.julialang.org/t/efficient-non-allocating-in-place-getindex-for-bitarray/42268

ImageGeoms.mask_or — Method

mask_or(mask)

compress 3D mask to 2D by logical or along z direction

ImageGeoms.maskit — Method

maskit(x::AbstractArray{<:Number}, mask)

opposite of embed

ImageGeoms.oversample — Method

ig_over = oversample(ig::ImageGeom, over::Int or NTuple)

Over-sample an image geometry by the factor over.

ImageGeoms.plot — Method

plot(ig, how ; kwargs...)

The how argument should be MIRTjim.jim to be useful.

ImageGeoms.upsample! — Method

upsample!(y, x, up::NTuple{Int})

Non-allocating version of y = upsample(x, up). Requires size(y) = ceil(size(x) * up).

ImageGeoms.upsample — Method

y = upsample(x, up::NTuple{Int})

Up-sample a Bool array x by "repeating" in all dimensions. Output size will be size(x) * up.